Move FreshIdGenerator to a separate file
